requirements.txt
langchain
langchain-groq
langchain_community
beautifulsoup4
chromadb
uuid
streamlit
python-dotenv